refactor(loading): extract shared LoadingProgress; show full steps in ProjectsView (#83)
The stepped progress list + spinner + repo-header column was duplicated between
LoadingOverlay (cold boot) and ProjectsView (in-app switches), and ProjectsView
only rendered a single status line. Extracted a shared <LoadingProgress> (one
definition of the phase->step state mapping) that both compose, so the Projects
page clone now shows the full step list too. Also decouples the loading status
line from the .text-card-title composite (its own .loading-status), fixing the
oversized status the .loading-card-scoped rule never reached.
