+++ app/src/views/ProjectsView/ProjectsView.tsx
@@ -17,8 +17,9 @@ import { GemIcon } from '@/components/GemIcon/GemIcon';
 import { PROJECTS_VIEW, type SourcePayload } from '@/state/stores/ui';
 import { SERVER_CONFIG } from '@/state/stores/serverConfig';
 import { SCAN_PROGRESS } from '@/state/stores/scanProgress';
-import { PENDING_SOURCE_LABEL, listRecents } from '@/state/stores/source';
-import { LOADING_STEP_LABELS, stepForPhase } from '@/constants/loadingSteps';
+import { listRecents } from '@/state/stores/source';
+import { stepForPhase } from '@/constants/loadingSteps';
+import { LoadingProgress } from '@/components/LoadingProgress/LoadingProgress';
 import { NewProjectForm } from '@/components/NewProjectForm/NewProjectForm';
 import { RecentsList } from '@/components/RecentsList/RecentsList';
 
@@ -94,15 +95,11 @@ export function ProjectsView({ onSubmit, onCancel, onClose }: ProjectsViewProps)
         <div class="landing-actions">
           {loading && scan ? (
             <section class="landing-card surface-sidebar">
-              <div class="landing-progress" role="status" aria-live="polite">
-                {PENDING_SOURCE_LABEL.value && (
-                  <div class="loading-pending-label">{PENDING_SOURCE_LABEL.value}</div>
-                )}
-                <div class="loading-spinner" />
-                <div class="text-card-title is-loading">
-                  {LOADING_STEP_LABELS[stepForPhase(scan.phase, scan.kind)]}
-                  {'…'}
-                </div>
+              <div class="landing-progress">
+                <LoadingProgress
+                  activeStep={stepForPhase(scan.phase, scan.kind)}
+                  kind={scan.kind}
+                />
                 <button type="button" class="btn-secondary" onClick={onCancel}>
                   Cancel
                 </button>
